Traktor: OSX fixes and an invalid assert removed from Proxy.
[git-p4: depot-paths = "//depot/traktor/": change = 1620]

+++ code/Drawing/Filters/QuantizeFilter.cpp
@@ -25,10 +25,10 @@ Ref< Image > QuantizeFilter::apply(const Image* image)
 		{
 			image->getPixel(x, y, in);
 
-			in.setRed(std::floorf(in.getRed() * m_steps + 0.5f) / m_steps);
-			in.setGreen(std::floorf(in.getGreen() * m_steps + 0.5f) / m_steps);
-			in.setBlue(std::floorf(in.getBlue() * m_steps + 0.5f) / m_steps);
-			in.setAlpha(std::floorf(in.getAlpha() * m_steps + 0.5f) / m_steps);
+			in.setRed(std::floor(in.getRed() * m_steps + 0.5f) / m_steps);
+			in.setGreen(std::floor(in.getGreen() * m_steps + 0.5f) / m_steps);
+			in.setBlue(std::floor(in.getBlue() * m_steps + 0.5f) / m_steps);
+			in.setAlpha(std::floor(in.getAlpha() * m_steps + 0.5f) / m_steps);
 
 			final->setPixel(x, y, in);
 		}
